bin/ext-accelerator/b2bpunchout/src/de/hybris/platform/b2b/punchout/actions/AuthenticationCheckPunchOutProcessingAction.java
/*
 * [y] hybris Platform
 *
 * Copyright (c) 2000-2015 hybris AG
 * All rights reserved.
 *
 * This software is the confidential and proprietary information of hybris
 * ("Confidential Information"). You shall not disclose such Confidential
 * Information and shall use it only in accordance with the terms of the
 * license agreement you entered into with hybris.
 *
 *  
 */
package de.hybris.platform.b2b.punchout.actions;

import org.cxml.CXML;


/**
 * This class will perform an authentication coming from the punch out system. The shared secret will be compared with
 * the configured shared secret. No cXML will be created, the only purpose is to authenticate the user.
 */
public class AuthenticationCheckPunchOutProcessingAction implements PunchOutProcessingAction<CXML, CXML>
{
	private AuthenticationVerifier punchoutAuthenticationVerifier;

	@Override
	public void process(final CXML input, final CXML output)
	{
		punchoutAuthenticationVerifier.verify(input);
	}

	public AuthenticationVerifier getPunchOutAuthenticationVerifier()
	{
		return punchoutAuthenticationVerifier;
	}

	public void setPunchOutAuthenticationVerifier(final AuthenticationVerifier punchoutAuthenticationVerifier)
	{
		this.punchoutAuthenticationVerifier = punchoutAuthenticationVerifier;
	}

}
